Patch review notes. I hope, the commented-out code with m_NCWin is still there for a reason - so I'm leaving the variable itself in commented too.

+++ src/shl.cpp
@@ -52,12 +52,12 @@ namespace SHL
 	{
 		if ( s >= end ) { return false; }
 
-		long l = end - s;
+		size_t l = end - s;
 		char buf[0x100];
 		std::vector<char> buf2;
 		char* p;
 
-		if ( l >= (int) sizeof( buf ) )
+		if ( l >= sizeof( buf ) )
 		{
 			buf2.resize( l + 1 );
 			p = buf2.data();
